:package: :: move ColorTheme.kt to color pakage
